Modified scroll update to keep up with Nomolos regardless of speed.
Modified scroll update to adjust Nomolos's screen coordinates based on difference with middle of screen. This looks smoother..no "jumping ahead."

Modified Nomolos' state logic so that his animation is updated while a button is pressed, and reset when no buttons are pressed. Added some constants to help with remembering how I'm using the bits in the state variable.

git-svn-id: svn://localhost/Nomolos/trunk@20 d887df96-dd9d-fa4b-9048-eb4ea8a40bc6