Adding logic for displaying the fight animation. This required some updates to how we update animation and when and so forth. For example now we must check for an on to off transition for the left and right button to reset the animation.
I notice when I reset animations I must set the frame count to 1 and then it plays the second frame first. This forced me to reverse the frames in the fight animations. This should definitely be fixed. I should be able to set everything to 0 and play an animation from the beginning.

git-svn-id: svn://localhost/Nomolos/trunk@123 d887df96-dd9d-fa4b-9048-eb4ea8a40bc6