Changing the hack to deal with the garbage tiles in Thoguth's arena to drawing a little patch of red bricks into the column buffer so the ppu upload routine won't overwrite them with black tiles.
Modifying Thoguth's arena to have both rooms duplicated. This is so we can decode both rooms into both nametables ahead of time, and when the scroll is re-enabled after the battle, the camera can "pop" to the right and not cause any garbage tiles to show up. This also precipitated changing the above hack, because changing what nametable to write to allowed garbage tiles to be written to the offscreen nametable, which *now* we DO want to preserve in order to allow camera popping to occur without issue.

Fascinating how such a small project can have so many consequences when you introduce new behavior or strange hacks!!

git-svn-id: svn://localhost/Nomolos/trunk@642 d887df96-dd9d-fa4b-9048-eb4ea8a40bc6