Adding ability for slides to load sprite chr data and a sprite overlay. Adding new sprite overlay data for slide 1. Adding new graphics and sprite graphics for slide 1. Expanding the Nomolos configuration to be 512kb to make room for all the new cut-scene graphics.
As of this commit, displaying a sprite overlay with the normal metasprite routine is broken. This is because an overlay metasprite has more entries than can be addressed by a single byte offset. We will have to write a new routine specialized for displaying the overlay. I tried this evening, but didn't quite make it.

git-svn-id: svn://localhost/Nomolos/branches/InfiniteNesLivesCart@1123 d887df96-dd9d-fa4b-9048-eb4ea8a40bc6